requirements.txt
appdirs==1.4.4
asgiref==3.8.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachelib==0.12.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
fake-useragent==1.5.1
Flask==3.0.2
Flask-Session==0.7.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.0
parse==1.20.1
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyee==11.1.0
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
requests==2.31.0
requests-html==0.10.0
rsa==4.9
soupsieve==2.5
SQLAlchemy==2.0.29
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
Werkzeug==3.0.1
zipp==3.18.1
